+++ frontend/src/pages/SubProjects/reducer.js
@@ -32,7 +32,9 @@ import {
   SUBPROJECT_SEARCH_TERM,
   SUBPROJECT_STORE_FILTERED_PROJECTS,
   SUBPROJECT_STORE_HIGHLIGHTING_REGEX,
-  SUBPROJECT_STORE_SEARCH_TERMS_AS_ARRAY
+  SUBPROJECT_STORE_SEARCH_TERMS_AS_ARRAY,
+  FETCH_FIRST_PROJECT_HISTORY_PAGE,
+  FETCH_FIRST_PROJECT_HISTORY_PAGE_SUCCESS
 } from "./actions";
 import { convertToURLQuery } from "../../helper";
 
@@ -173,6 +175,7 @@ export default function detailviewReducer(state = defaultState, action) {
       return state.set("showProjectAssignees", false);
     case HIDE_SUBPROJECT_ADDITIONAL_DATA:
       return state.set("isSubProjectAdditionalDataShown", false);
+    case FETCH_FIRST_PROJECT_HISTORY_PAGE:
     case FETCH_NEXT_PROJECT_HISTORY_PAGE:
       return state.set("isHistoryLoading", true);
     case SET_TOTAL_PROJECT_HISTORY_ITEM_COUNT:
@@ -180,10 +183,15 @@ export default function detailviewReducer(state = defaultState, action) {
         totalHistoryItemCount: action.totalHistoryItemsCount,
         lastHistoryPage: action.lastHistoryPage
       });
-
     case FETCH_NEXT_PROJECT_HISTORY_PAGE_SUCCESS:
       return state.merge({
         historyItems: state.get("historyItems").concat(fromJS(action.events).reverse()),
+        currentHistoryPage: action.currentHistoryPage,
+        isHistoryLoading: false
+      });
+    case FETCH_FIRST_PROJECT_HISTORY_PAGE_SUCCESS:
+      return state.merge({
+        historyItems: fromJS(action.events).reverse(),
         currentHistoryPage: action.currentHistoryPage,
         isHistoryLoading: false
       });